Nuzcraft patch 1 (#10)
* Player Greatsword Character

Rexpaint and png of the player character with a greatsword

* Update character storage and draw function

* Update character storage and draw function

* Update character storage and draw function

* Rename Functions to Functions.py

* Update character storage and draw function

* Basic player movement.

I changed the way the object 'character' is grabbed and drawn to the screen. I also added a clear function.

* Basic player movement.

I added in logic for the player to move, and some key bindings as well.

* Basic player movement.

I added in some keybindings and placed the player characters into a dictionary, to make it easier to find them.

* Basic player movement

This file will hold all the key and mouse interactions.